Create Cloud Data Base
This project is a Python 3 Cloud Data base. It is a backend mulithreading TCP client/server program. Executed in the CMD window, the end user and put full name, address, phone number, and customer picture that allows multiple features for each of these inputs. features such as downloading between users uploaded photos and full databases. Users can also upload databases and there is an included menu for the user to keep track of the commands available to them.